Really first draft of object composition
