+++ lib/client/services/commons/index.js
@@ -1,0 +1,2 @@
+export { default as base } from './base'
+export { default as query } from './query'
